requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==14.5.14
Django==1.7.3
django-activity-stream==0.5.1
django-allauth==0.19.0
django-angular==0.7.10
django-autoslug==1.7.2
django-bootstrap3==5.1.1
django-cities-light==3.0.4
django-filter==0.9.2
django-guardian==1.2.5
djangorestframework==3.0.4
Markdown==2.5.2
oauthlib==0.7.2
python-openid==2.2.5
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
tornado==4.0.2
Unidecode==0.4.17
